feat(spectra-api): smuggling guard, worker-pool offload, TLS gateway with HTTP/2, db pool/migrations and JSON DOM hosts
- HTTP/1.1 parser rejects Transfer-Encoding + Content-Length together
  (RFC 9112) as ConflictingFraming -> 400; divergent duplicates covered
- sync handlers run on a bounded worker pool (per-connection awaiting
  state, catch_unwind -> 500); the event loop never executes handlers
  inline anymore — slow handler no longer freezes all connections
- TLS gateway: when certificates are set a tokio listener terminates TLS
  with ALPN fan-out (h2 -> existing pipeline, http/1.1 -> same dispatcher);
  new hosts server.set_tls_certificate/tls_local_port
- db pool + migrations exposed: pool.sqlite_open/close/with_connection
  (lease returned by sqlite.close), migrate.apply_sqlite/status_sqlite on
  top of spectra-db's checksummed migrator
- JSON DOM: parse/value_kind/len/get/at/text/number_bits/bool/free/
  stringify over a handle store (serde_json-backed)
- string readers flipped to packed UTF-8 byte buffers (api_registration,
  db_core, health, trace)

+++ packages/spectra-api/src/health.rs
@@ -115,7 +115,7 @@ unsafe fn string(value: i64) -> Option<String> {
     if value == 0 {
         return None;
     }
-    let ptr = value as *const i64;
+    let ptr = value as *const u8;
     let mut bytes = Vec::new();
     for index in 0..4096 {
         let byte = *ptr.add(index) as u8;
